+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -11,7 +11,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 180;
         int actual = service.sumSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -23,7 +23,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 15;
         int actual = service.averageSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -35,7 +35,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 8;
         int actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -47,7 +47,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 9;
         int actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -59,7 +59,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         int actual = service.lessAverageSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -71,7 +71,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
 
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         int actual = service.moreAverageSales(sales);
         Assertions.assertEquals(expected, actual);
